+++ src/accounts/views.py
@@ -1,12 +1,11 @@
 from __future__ import unicode_literals
-from django.core.urlresolvers import reverse_lazy
-from django.views import generic
-from django.contrib.auth import get_user_model
-from django.contrib import auth
-from django.contrib import messages
+
 from authtools import views as authviews
 from braces import views as bracesviews
-from django.conf import settings
+from django.contrib import messages
+from django.contrib.auth import get_user_model
+from django.core.urlresolvers import reverse_lazy
+
 from . import forms
 
 User = get_user_model()
@@ -19,16 +18,18 @@ class LoginView(bracesviews.AnonymousRequiredMixin,
 
     def form_valid(self, form):
         redirect = super(LoginView, self).form_valid(form)
-        remember_me = form.cleaned_data.get('remember_me')
-        if remember_me is True:
-            ONE_MONTH = 30*24*60*60
-            expiry = getattr(settings, "KEEP_LOGGED_DURATION", ONE_MONTH)
-            self.request.session.set_expiry(expiry)
+        # remember_me = form.cleaned_data.get('remember_me')
+        # if remember_me is True:
+        #     ONE_MONTH = 30*24*60*60
+        #     expiry = getattr(settings, "KEEP_LOGGED_DURATION", ONE_MONTH)
+        #     self.request.session.set_expiry(expiry)
         return redirect
 
 
 class LogoutView(authviews.LogoutView):
     url = reverse_lazy('home')
+
+
 #
 #
 # class SignUpView(bracesviews.AnonymousRequiredMixin,