+++ src/action/apps.py
@@ -6,4 +6,4 @@ from django.apps import AppConfig
 
 class ActionConfig(AppConfig):
     name = 'action'
-    verbose_name = 'Action, Condition & Filter Management'
+    verbose_name = 'Actions, Conditions, Filters, Emails'