+++ src/email_action/__init__.py
@@ -1,3 +1,0 @@
-from __future__ import unicode_literals, print_function
-
-default_app_config = 'email_action.apps.EmailActionConfig'