+++ src/email_action/apps.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-from __future__ import unicode_literals, print_function
-
-from django.apps import AppConfig
-
-
-class EmailActionConfig(AppConfig):
-    name = 'email_action'
-    verbose_name = 'Action Execution: Send Emails'