+++ src/logs/ops.py
@@ -20,6 +20,7 @@ log_types = {
     'column_add': 'Column added',
     'column_rename': 'Column renamed',
     'column_delete': 'Column deleted',
+    'column_clone': 'Column cloned',
     'action_create': 'Action created',
     'action_update': 'Action updated',
     'action_delete': 'Action deleted',